feat(chat-panel): add new components for brainstorming and chat states, enhance document reference handling
