+++ apps/DocFlow/src/utils/editor/index.ts
@@ -3,6 +3,7 @@
  */
 
 export * from './get-render-container';
+export * from './get-selection-line-range';
 export * from './is-custom-node-selected';
 export * from './is-text-selected';
 export * from './helpers';
